feat(nvim): update plugin versions and add new plugins
This commit updates the versions of various plugins used in the Neovim
configuration. It also adds new plugins for enhanced functionality. The
added plugins include 'mini-surround', 'git', and 'sql'. The plugin
versions are updated in the 'lazy-lock.json' file and the new plugins
are added in the 'lazyvim.json' file.