feat(qtile config): add change_sink function on screen change
This commit introduces a new function `change_sink` that is triggered
on screen change. It checks the output of `pactl list sinks short` and
sets the default sink to the first HDMI stereo sink that is not
currently running. This is useful for automatically switching audio
output when a new screen (with potential audio capabilities) is
connected.