refactor: update plugin versions and clean up scripts
This commit updates the versions of various plugins in the
lazy-lock.json file. It also cleans up some scripts, removing
unnecessary lines and improving the formatting. The changes in the
scripts include using the $() syntax for command substitution in
link.sh and removing the xidlehook service start command in
qtile/autostart.sh.