feat: Add PactlVolume widget to qtile config
This commit introduces a new widget, PactlVolume, to the qtile
configuration. This widget is designed to control the system volume
using pactl commands. The volume control keys are also updated to use
this new widget. The widget is added to the toolbar, and it supports
volume increase, decrease, and mute functionalities.

A new file, pactl_volume.py, is created to define the PactlVolume
class. This class extends the Volume class from libqtile.widget.volume
and overrides several methods to use pactl commands for volume control.